Fix frequency cap. Switch to long

+++ src-ui/sdr/hackrf.cpp
@@ -105,7 +105,7 @@ void SDRHackRF::drawUI()
     ImGui::End();
 }
 
-void SDRHackRF::setFrequency(int frequency)
+void SDRHackRF::setFrequency(long frequency)
 {
     d_frequency = frequency;
     std::memcpy(this->frequency, std::to_string((float)d_frequency / 1e6).c_str(), std::to_string((float)d_frequency / 1e6).length());
