+++ src-ui/sdr/hackrf.h
@@ -24,7 +24,7 @@ public:
     void stop();
     void drawUI();
     static void init();
-    virtual void setFrequency(int frequency);
+    virtual void setFrequency(long frequency);
     static std::vector<std::tuple<std::string, sdr_device_type, uint64_t>> getDevices();
 };
 #endif
